+++ app/src/main/java/com/babylon/wallet/android/presentation/nfc/NfcNavGraph.kt
@@ -27,4 +27,3 @@ fun NavGraphBuilder.nfcDialog(
         )
     }
 }
-
