+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/debug/ArculusToolsScreen.kt
@@ -55,4 +55,3 @@ private fun ArculusToolsScreen(
         }
     }
 }
-
