+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/debug/ArculusToolsViewModel.kt
@@ -33,4 +33,3 @@ class ArculusToolsViewModel @Inject constructor(
         }
     }
 }
-
