+++ app/src/main/java/com/babylon/wallet/android/utils/NavigationExtensions.kt
@@ -1,0 +1,13 @@
+package com.babylon.wallet.android.utils
+
+import androidx.navigation.NavController
+
+@Suppress("SwallowedException")
+fun NavController.routeExist(route: String): Boolean {
+    return try {
+        getBackStackEntry(route)
+        true
+    } catch (e: Exception) {
+        false
+    }
+}
